Tooth based time to angle coversion is only used by a few decoders. So move the functions into decoders.cpp
