+++ speeduino/decoders.cpp
@@ -293,6 +293,47 @@ void loggerTertiaryISR(void)
   }  
 }
 
+#if false
+#if !defined(UNIT_TEST)
+static
+#endif
+uint32_t angleToTimeIntervalTooth(uint16_t angle) {
+  noInterrupts();
+  if(BIT_CHECK(decoderState, BIT_DECODER_TOOTH_ANG_CORRECT))
+  {
+    unsigned long toothTime = (toothLastToothTime - toothLastMinusOneToothTime);
+    uint16_t tempTriggerToothAngle = triggerToothAngle; // triggerToothAngle is set by interrupts
+    interrupts();
+    
+    return (toothTime * (uint32_t)angle) / tempTriggerToothAngle;
+  }
+  //Safety check. This can occur if the last tooth seen was outside the normal pattern etc
+  else { 
+    interrupts();
+    return angleToTimeMicroSecPerDegree(angle); 
+  }
+}
+#endif
+
+static uint16_t timeToAngleIntervalTooth(uint32_t time)
+{
+    noInterrupts();
+    //Still uses a last interval method (ie retrospective), but bases the interval on the gap between the 2 most recent teeth rather than the last full revolution
+    if(BIT_CHECK(decoderState, BIT_DECODER_TOOTH_ANG_CORRECT))
+    {
+      unsigned long toothTime = (toothLastToothTime - toothLastMinusOneToothTime);
+      uint16_t tempTriggerToothAngle = triggerToothAngle; // triggerToothAngle is set by interrupts
+      interrupts();
+
+      return (unsigned long)(time * (uint32_t)tempTriggerToothAngle) / toothTime;
+    }
+    else { 
+      interrupts();
+      //Safety check. This can occur if the last tooth seen was outside the normal pattern etc
+      return timeToAngleDegPerMicroSec(time);
+    }
+}
+
 static inline bool IsCranking(const statuses &status) {
   return (status.RPM < status.crankRPM) && (status.startRevolutions == 0U);
 }
